refactoring to remove duplicate code and restructure the terminalView to take props
