+++ my-app/app/page.tsx
@@ -1,15 +1,55 @@
 "use client";
-import { useState } from "react";
+
 import ChangeDot from "./MyComponents/ChangeDot";
 import CommandBox from "./MyComponents/CommandBox";
 import TerminalView from "./MyComponents/TerminalView";
 import Title from "./MyComponents/Title";
 import type { CommitType } from "@/types";
+import { useGlobalStore } from "@/stores/globalStorage";
+import { useEffect, useState } from "react";
 import { toast } from "sonner";
 
 export default function Home() {
   const [commits, setCommits] = useState<CommitType[]>([]);
-  const branches: string[] = [];
+  const [terminalOutput, setTerminalOutput] = useState<JSX.Element[]>([]);
+  // const branchesArr: string[] = [];
+
+  const { currentBranch, branches } = useGlobalStore();
+  const [unstagedCommits, setUnstagedCommits] = useState<CommitType[]>([]);
+
+  function getUnstagedCommits() {
+    const commits = branches.get(currentBranch)?.commits;
+    if (commits === undefined) {
+      // branches has yet to have any commits to it
+      return setUnstagedCommits([] as CommitType[]);
+    } else {
+      // if c.stagedChange == undefined/false will be filtered
+      setUnstagedCommits(commits.filter((c) => c.state != "commited"));
+    }
+  }
+
+  useEffect(() => {
+    getUnstagedCommits();
+  }, [currentBranch, branches]);
+
+  function displayCommit(commit: CommitType) {
+    return (
+      <span
+        className={
+          "flex flex-row gap-2" + commit.state == "unstaged"
+            ? "text-red-500"
+            : "text-green-400"
+        }
+      >
+        <p>
+          {"<"}
+          {commit.commitId}
+          {">"}
+        </p>
+        <p>{commit.commitMessage.slice(8)}...</p>
+      </span>
+    );
+  }
   function handleCreateCommit(
     commitMessage: string,
     commitId: number,
@@ -36,17 +76,15 @@ export default function Home() {
       });
       return;
     }
-    if(commandComponents[1] == "branch" ){
-
+    if (commandComponents[1] == "branch") {
     }
-  
   }
 
   return (
     <div className="bg-gray-800 h-screen">
       <div className="mr-60 ml-60">
         <Title />
-        <TerminalView terminalOutput="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi "/>
+        <TerminalView terminalOutput={terminalOutput} />
         <CommandBox
           createCommit={handleCreateCommit}
           handleGitCommit={handleGitCommand}
